Fixed again type for std::min.

+++ stellarsolver/internalsextractorsolver.cpp
@@ -249,8 +249,8 @@ int InternalSextractorSolver::runSEPSextractor()
         // Limit it to PARTITION_THREADS
         if ( (w * h) / (W_PARTITION_SIZE * H_PARTITION_SIZE) > m_PartitionThreads)
         {
-            W_PARTITION_SIZE = w / std::min(m_PartitionThreads / 2, 1UL);
-            H_PARTITION_SIZE = h / std::min(m_PartitionThreads / 4, 1UL);
+            W_PARTITION_SIZE = w / std::min(m_PartitionThreads / 2, (uint32_t)1);
+            H_PARTITION_SIZE = h / std::min(m_PartitionThreads / 4, (uint32_t)1);
         }
 
         int horizontalPartitions = w / W_PARTITION_SIZE;
